modify proj tree
