+++ examples/rpclookup/lookup_client/msghandler.go
@@ -4,12 +4,22 @@ import (
 	"io"
 	"net"
 	"time"
-	"github.com/wondywang/rpclookup/core/util"
+	"github.com/ikingfisher/jellyfish/core/util"
+	"github.com/ikingfisher/jellyfish/core/codec"
 )
 
-func HeartBeatWrite(conn net.Conn, done chan int) {
-	body := "hello"
-	buf := []byte("H")
+func HandleMsgWrite(conn net.Conn, done chan int) {
+	// body := "hello"
+	buf := []byte("D")
+	var req codec.Request
+	req.Cmd = "QueryList"
+	req.Body = append(req.Body, string("request from client")...)
+	body, err := codec.ReqEncode(req)
+	if err != nil {
+		logger.Error("request encode failed! %s", err.Error())
+		return
+	}
+
 	bodySize := util.Int64ToBytes(int64(len(body)))
 	nanoTime := time.Now().UnixNano()
 	seq := util.Int64ToBytes(nanoTime)
@@ -18,7 +28,8 @@ func HeartBeatWrite(conn net.Conn, done chan int) {
 	buf = append(buf, bodySize...)
 	buf = append(buf, seq...)
 	buf = append(buf, body...)
-	_, err := conn.Write(buf)
+
+	_, err = conn.Write(buf)
 	if err != nil {
 		if opErr, ok := err.(*net.OpError); ok && opErr.Timeout() {
 			logger.Debug("write message timeout. %s", err.Error())
@@ -30,7 +41,7 @@ func HeartBeatWrite(conn net.Conn, done chan int) {
 	}
 }
 
-func HeartBeatRead(conn net.Conn, done chan int) {
+func HandleMsgRead(conn net.Conn, done chan int) {
 	buf := make([]byte, 1024)
 	len, err := conn.Read(buf)
 	if err != nil {
