+++ lookupd/lookupd.go
@@ -1,9 +1,9 @@
 package lookupd
 
 import (
-	"github.com/wondywang/rpclookup/core/client"
-	"github.com/wondywang/rpclookup/core/lg"
-	"github.com/wondywang/rpclookup/core/util"
+	"github.com/ikingfisher/jellyfish/core/client"
+	"github.com/ikingfisher/jellyfish/core/lg"
+	"github.com/ikingfisher/jellyfish/core/util"
 	"os"
 	"net"
 	"io"
