+++ make.include
@@ -1,7 +1,7 @@
 ########################################
 #	Common Makefile for Include
 ########################################
-patterns = "/rpclookup|/outerlib"
+patterns = "/jellyfish|/outerlib"
 PROJ_ROOT = $(shell pwd | awk -F$(patterns) '{print $$1}')
 
-GOPATH := $(PROJ_ROOT)/rpclookup:$(GOPATH)
+GOPATH := $(PROJ_ROOT)/jellyfish:$(GOPATH)
